Moving more functions out of DefaultDelivery to be used by both DefaultDelivery and OkHttpDelivery
