+++ bugsnag-android-core/src/test/java/com/bugsnag/android/DeliveryTest.kt
@@ -1,5 +1,6 @@
 package com.bugsnag.android
 
+import com.bugsnag.android.internal.DeliveryHelper.getDeliveryStatus
 import org.junit.Assert.assertEquals
 import org.junit.Test
 
@@ -7,17 +8,16 @@ class DeliveryTest {
 
     @Test
     fun testResponseCodeMapping() {
-        val delivery = DefaultDelivery(null, "myApiKey", 10000, NoopLogger)
-        assertEquals(DeliveryStatus.DELIVERED, delivery.getDeliveryStatus(202))
-        assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(503))
-        assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(0))
-        assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(408))
-        assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(429))
-        assertEquals(DeliveryStatus.FAILURE, delivery.getDeliveryStatus(400))
-        assertEquals(DeliveryStatus.FAILURE, delivery.getDeliveryStatus(401))
-        assertEquals(DeliveryStatus.FAILURE, delivery.getDeliveryStatus(498))
-        assertEquals(DeliveryStatus.FAILURE, delivery.getDeliveryStatus(499))
-        assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(408))
-        assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(429))
+        assertEquals(DeliveryStatus.DELIVERED, getDeliveryStatus(202))
+        assertEquals(DeliveryStatus.UNDELIVERED, getDeliveryStatus(503))
+        assertEquals(DeliveryStatus.UNDELIVERED, getDeliveryStatus(0))
+        assertEquals(DeliveryStatus.UNDELIVERED, getDeliveryStatus(408))
+        assertEquals(DeliveryStatus.UNDELIVERED, getDeliveryStatus(429))
+        assertEquals(DeliveryStatus.FAILURE, getDeliveryStatus(400))
+        assertEquals(DeliveryStatus.FAILURE, getDeliveryStatus(401))
+        assertEquals(DeliveryStatus.FAILURE, getDeliveryStatus(498))
+        assertEquals(DeliveryStatus.FAILURE, getDeliveryStatus(499))
+        assertEquals(DeliveryStatus.UNDELIVERED, getDeliveryStatus(408))
+        assertEquals(DeliveryStatus.UNDELIVERED, getDeliveryStatus(429))
     }
 }
