+++ bugsnag-plugin-android-okhttp/src/main/java/com/bugsnag/android/okhttp/OkHttpDelivery.kt
@@ -8,8 +8,10 @@ import com.bugsnag.android.DeliveryStatus
 import com.bugsnag.android.EventPayload
 import com.bugsnag.android.Logger
 import com.bugsnag.android.Session
+import com.bugsnag.android.internal.DeliveryHelper.getDeliveryStatus
+import com.bugsnag.android.internal.DeliveryHelper.retrieveSha1Digest
+import com.bugsnag.android.internal.DeliveryHelper.serializePayload
 import com.bugsnag.android.internal.JsonHelper
-import com.bugsnag.android.internal.serializePayload
 import okhttp3.Headers
 import okhttp3.Headers.Companion.toHeaders
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
@@ -58,11 +60,15 @@ class OkHttpDelivery constructor(
         return try {
             val mediaType = "application/json".toMediaTypeOrNull()
             val headersNotNull = convertHeaders(headers)
-            val request = Request.Builder()
-                .headers(headersNotNull.toHeaders()) // add error API headers
-                .url(url) // use error API endpoint
+            val requestSetUp = Request.Builder()
+                .headers(headersNotNull.toHeaders())
+                .url(url)
                 .post(json.toRequestBody(mediaType))
-                .build()
+
+            // calculate the SHA-1 digest and add to headers
+            retrieveSha1Digest(json)?.let { requestSetUp.addHeader("Bugsnag-Integrity", it) }
+
+            val request = requestSetUp.build()
             val response = existingOkHttpClient.newCall(request).execute()
             logRequestInfo(response.code, response.message, response.headers)
             getDeliveryStatus(response.code)
@@ -91,19 +97,4 @@ class OkHttpDelivery constructor(
     private fun convertHeaders(headers: Map<String, String?>): Map<String, String> {
         return headers.filterValues { it != null }.mapValues { it.value as String }
     }
-
-    @Suppress("MagicNumber")
-    internal fun getDeliveryStatus(responseCode: Int): DeliveryStatus {
-        return when {
-            responseCode in 200..299 -> DeliveryStatus.DELIVERED
-            isUnrecoverableStatusCode(responseCode) -> DeliveryStatus.FAILURE
-            else -> DeliveryStatus.UNDELIVERED
-        }
-    }
-
-    @Suppress("MagicNumber")
-    private fun isUnrecoverableStatusCode(responseCode: Int) =
-        responseCode in 400..499 && // 400-499 are considered unrecoverable
-            responseCode != 408 && // except for 408
-            responseCode != 429 // and 429
 }
